Porting previous commit to new elastic-search head structure after rebase.
Adding the ability to delete records via the _head UI. This can now bee done by right clicking a row in the
Browser, or when viewing a table of results, and select 'Delete Recordw' to delete the row.
The row will be deleted from elasticsearch, and then the table.
This further improves the ability for non-technical users to use the _head plug-in for QA or other purposes.

In order to have do this, I added a right click handler to jsacx-widgets.js. Only if the event is intercepted
and .preventDefault() is called on the event will the regular browser context menu not be shown.
