--- lib/jsacx/src/jsacx-widgets.css
@@ -1,324 +1,0 @@
-.ui-left { float: left; }
-.ui-right { float: right; }
-.require { color: #a00; }
-
-.uiToolbar {
-	height: 28px;
-	background: #fdfefe;
-	background: -moz-linear-gradient(top, #fdfefe, #eaedef);
-	background: -webkit-linear-gradient(top, #fdfefe, #eaedef);
-	border-bottom: 1px solid #d2d5d7;
-	padding: 3px 10px;
-}
-
-.uiToolbar H2 {
-	display: inline-block;
-	font-size: 120%;
-	margin: 0;
-	padding: 5px 20px 5px 0;
-}
-
-.uiButton {
-	padding: 0;
-	border: 0;
-	margin: 3px;
-	width: auto;
-	overflow: visible;
-	cursor: pointer;
-	background: transparent;
-}
-
-.uiButton-content {
-	height: 20px;
-	min-width: 73px;
-	border: 1px solid #668dc6;
-	border-radius: 2px;
-	-moz-border-radius: 2px;
-	-webkit-border-radius: 2px;
-	background: #96c6eb;
-	background: -moz-linear-gradient(top, #96c6eb, #5296c7);
-	background: -webkit-linear-gradient(top, #96c6eb, #5296c7);
-	color: white;
-	font-weight: bold;
-}
-
-.moz .uiButton-content { margin: 0 -2px; }
-
-.uiButton-label {
-    padding: 2px 6px;
-    white-space: nowrap;
-}
-BUTTON:hover .uiButton-content {
-	background: #2777ba;
-	background: -moz-linear-gradient(top, #6aaadf, #2777ba);
-	background: -webkit-linear-gradient(top, #6aaadf, #2777ba);
-}
-BUTTON.active .uiButton-content,
-BUTTON:active .uiButton-content {
-	background: #2575b7;
-	background: -moz-linear-gradient(top, #2576b8, #2575b7);
-	background: -webkit-linear-gradient(top, #2576b8, #2575b7);
-}
-BUTTON.disabled .uiButton-content,
-BUTTON.disabled:active .uiButton-content {
-    border-color: #c6c6c6;
-    color: #999999;
-    background: #ddd;
-    background: -moz-linear-gradient(top, #ddd, #ddd);
-    background: -webkit-linear-gradient(top, #ddd, #ddd);
-}
-
-BUTTON.disabled {
-    cursor: default;
-}
-
-BUTTON::-moz-focus-inner {
-    border: none;
-}
-
-/* acx.ui.MenuButton */
-.uiMenuButton {
-	display: inline-block;
-}
-
-.uiMenuButton.active .uiButton-content,
-.uiMenuButton:active .uiButton-content {
-	-webkit-border-bottom-right-radius: 0px;
-	-webkit-border-bottom-left-radius: 0px;
-	-moz-border-radius-bottomright: 0px;
-	-moz-border-radius-bottomleft: 0px;
-	border-bottom-right-radius: 0px;
-	border-bottom-left-radius: 0px;
-}
-
-.uiMenuButton .uiButton-content {
-	padding-right: 3px;
-	border-radius: 8px;
-	-moz-border-radius: 8px;
-	-webkit-border-radius: 8px;
-	height: auto;
-	min-width: 0;
-}
-.uiMenuButton .uiButton-label {
-	background-image: url('images/menu-button.gif');
-	background-position: right 50%;
-	background-repeat: no-repeat;
-	padding: 0px 17px 0px 5px;
-	text-align: left;
-}
-
-/* acx.ui.SplitButton */
-
-.uiSplitButton {
-	white-space: nowrap;
-}
-
-.uiSplitButton .uiButton:first-child {
-	margin-right: 0;
-	display: inline-block;
-}
-
-.uiSplitButton .uiButton:first-child .uiButton-content {
-	border-right-width: 1;
-	border-right-color: #5296c7;
-	border-top-right-radius: 0;
-	border-bottom-right-radius: 0;
-	min-width: 0;
-}
-
-.uiSplitButton .uiMenuButton {
-	margin-left: 0;
-}
-
-.uiSplitButton .uiButton:last-child .uiButton-content {
-	border-radius: 2px;
-	border-left-width: 1;
-	border-left-color: #96c6eb;
-	border-top-left-radius: 0;
-	border-bottom-left-radius: 0;
-	height: 20px;
-}
-
-.uiSplitButton .uiButton:last-child .uiButton-label {
-	padding: 2px 17px 2px 6px;
-	margin-left: -8px;
-}
-.uiSplitMenuPanel .uiMenuPanel-label {
-	margin-left: 1em;
-	padding-left: 4px;
-}
-
-.uiSplitMenuPanel .uiMenuPanel-item.selected .uiMenuPanel-label:before {
-	content: "\2713";
-	width: 12px;
-	margin-left: -12px;
-	display: inline-block;
-
-}
-/* acx.ui.Table */
-
-.uiTable TABLE {
-	border-collapse: collapse;
-}
-
-.uiTable-body {
-	overflow-y: scroll;
-	overflow-x: auto;
-}
-
-.uiTable-headers {
-	overflow-x: hidden;
-}
-
-.uiTable-body TD {
-	white-space: nowrap;
-}
-
-.uiTable-body .uiTable-header-row TH,
-.uiTable-body .uiTable-header-row TH DIV {
-	padding-top: 0;
-	padding-bottom: 0;
-}
-
-.uiTable-body .uiTable-header-cell > DIV {
-	height: 0;
-	overflow: hidden;
-}
-
-.uiTable-headercell-menu {
-	float: right;
-}
-
-/* acx.ui.*Panel */
-#uiModal {
-	background: black;
-}
-
-.uiPanel {
-	-moz-box-shadow: -1px 2.5px 4px -3px black, -1px -2.5px 4px -3px black, 3px 2.5px 4px -3px black, 3px -2.5px 4px -3px black;
-	position: absolute;
-	background: #eee;
-	border: 1px solid #666;
-}
-
-.uiPanel-titleBar {
-	text-align: center;
-	font-weight: bold;
-	padding: 2px 0;
-	background: rgba(223, 223, 223, 0.75);
-	background: -moz-linear-gradient(top, rgba(223, 223, 223, 0.75), rgba(193, 193, 193, 0.75), rgba(223, 223, 223, 0.75));
-	background: -webkit-linear-gradient(top, rgba(223, 223, 223, 0.75), rgba(193, 193, 193, 0.75), rgba(223, 223, 223, 0.75));
-	border-bottom: 1px solid #bbb;
-}
-
-.uiPanel-close {
-	cursor: pointer;
-	border: 1px solid #aaa;
-	background: #fff;
-	color: #fff;
-	float: left;
-	height: 10px;
-	left: 3px;
-	line-height: 9px;
-	padding: 1px 0;
-	position: relative;
-	text-shadow: 0 0 1px #000;
-	top: 0px;
-	width: 12px;
-}
-.uiPanel-close:hover {
-	background: #eee;
-}
-
-.uiPanel-body {
-	overflow: auto;
-}
-
-.dark.uiPanel {
-	background: rgba(0, 0, 0, 0.75);
-	color: white;
-	-moz-border-radius: 8px;
-	-webkit-border-radius: 8px;
-	padding: 1px;
-}
-.dark .uiPanel-titleBar {
-	background: rgba(74, 74, 74, 0.75);
-	background: -moz-linear-gradient(top, rgba(84, 84, 84, 0.75), rgba(54, 54, 54, 0.75), rgba(64, 64, 64, 0.75));
-	background: -webkit-linear-gradient(top, rgba(84, 84, 84, 0.75), rgba(54, 54, 54, 0.75), rgba(64, 64, 64, 0.75));
-	-moz-border-radius: 8px 8px 0px 0px;
-	-webkit-border-top-right-radius: 8px;
-	-webkit-border-top-left-radius: 8px;
-	padding: 1px 0 2px 0;
-	border-bottom: 0;
-}
-.dark .uiPanel-close {
-	-moz-border-radius: 6px;
-	-webkit-border-radius: 6px;
-	height: 13px;
-	width: 13px;
-	background: #ccc;
-	left: 3px;
-	top: 1px;
-	color: #333;
-	text-shadow: #222 0 0 1px;
-	line-height: 11px;
-	border: 0;
-	padding: 0;
-}
-.dark .uiPanel-close:hover {
-	background: #eee;
-}
-
-.dark .uiPanel-body {
-	background: transparent;
-	padding: 20px;
-	-moz-border-radius: 0 0 8px 8px;
-	-webkit-border-bottom-right-radius: 8px;
-	-webkit-border-bottom-left-radius: 8px;
-	border: 1px solid #222;
-}
-
-.uiPanelForm-field {
-	display: block;
-	padding: 2px 0;
-	clear: both;
-}
-
-.uiPanelForm-label {
-	float: left;
-	width: 200px;
-	padding: 3px 7px;
-	text-align: right;
-}
-
-.uiMenuPanel {
-	border: 1px solid #668dc6;
-	position: absolute;
-	background: #96c6eb;
-	color: white;
-}
-
-.uiMenuPanel LI {
-	list-style: none;
-	border-bottom: 1px solid #668dc6;
-}
-
-.uiMenuPanel LI:hover {
-	background: #2575b7;
-}
-
-.uiMenuPanel LI:last-child {
-	border-bottom: 0;
-} 
-
-.uiMenuPanel-label {
-	white-space: nowrap;
-	padding: 2px 10px 2px 10px;
-	cursor: pointer;
-}
-
-.disabled .uiMenuPanel-label {
-	cursor: auto;
-	color: #888;
-}
-
