+++ src/app/ui/menuPanel/menuPanel.js
@@ -14,11 +14,16 @@
 		},
 		open: function(jEv) {
 			this._super(jEv);
-			var cx = this; setTimeout(function() { $(document).bind("click", cx._close_handler); }, 50);
+			var cx = this;
+			setTimeout(function() {
+				$(document).bind("click", cx._close_handler);
+				$(document).bind("contextmenu", cx._close_handler);
+			}, 50);
 		},
 		_close_handler: function(jEv) {
 			this._super(jEv);
 			$(document).unbind("click", this._close_handler);
+			$(document).unbind("contextmenu", this._close_handler);
 		},
 		_main_template: function() {
 			return { tag: "DIV", cls: this._baseCls, children: this.config.items.map(this._menuItem_template, this) };
