+++ src/app/ui/resultTable/resultTable.js
@@ -5,12 +5,15 @@
 	ui.ResultTable = ui.Table.extend({
 		defaults: {
 			width: 500,
-			height: 400
+			height: 400,
+			cluster: null  // (required) instanceof es.Cluster
 		},
 
 		init: function() {
 			this._super();
+			this.cluster = this.config.cluster
 			this.on("rowClick", this._showPreview_handler);
+			this.on("rowRightClick", this._showDeleteMenu_handler);
 			this.selectedRow = null;
 			$(document).bind("keydown", this._nav_handler);
 		},
@@ -49,6 +52,48 @@
 		},
 		_showPreview_handler: function(obj, data) {
 			this.showPreview(this.selectedRow = data.row);
+		},
+		_showDeleteMenu_handler: function(obj, data) {
+			var menuElement = $({tag: "DIV", cls: "ui"});
+			var menuPanel = new ui.MenuPanelOnElement({
+				element: data.row,
+				items: [
+					{ text: i18n.text("Browser.DeleteRecord"), onclick: function() { this._deleteRow(data.row) }.bind(this) }
+				]
+			});
+			var menuPaneljQ = $(menuPanel).appendTo(menuElement);
+			menuPanel.open(data.event);
+			menuPaneljQ.offset({
+				top: data.event.pageY,
+				left: data.event.pageX
+			});
+			data.event.preventDefault();
+		},
+		_deleteRow: function(row) {
+			var rowData = row.data("row");
+			rowData = rowData["_source"];
+			// Delete from the index.
+			this.cluster.delete(rowData["_index"] + "/" + rowData["_type"] + "/" + encodeURIComponent(rowData["_id"]));
+			// Remove the row from table.
+			row.remove();
+		}
+	});
+
+	ui.MenuPanelOnElement = ui.MenuPanel.extend({
+		defaults: {
+			element: null // Required DOM element
+		},
+		init: function() {
+			this._super();
+			this.element = $(this.config.element);
+		},
+		open : function(jEv) {
+			this.element.addClass("selected");
+			this._super(jEv);
+		},
+		_close_handler: function(jEv) {
+			this.element.removeClass("selected");
+			this._super(jEv);
 		}
 	});
